// connect247new/mssqlwebclient/src/app/models/databaseInstance.ts
// export class DatabaseInstance {

//   MachineName: string;
//   dbName: string
// }

export class DatabaseInstance {

  "machineName": string;
        "instanceName": string;
        "fullInstanceName": string;
        "productversion": string;
        "productlevel": string;
        "computerNamePhysicalNetBIOS": string;
        "edition": string;
        "isFullTextInstalled": number;
        "isClustered": number;
        "isIntegratedSecurityOnly": number;
        "totalDatabases": number;
        "dbName": string;
        "dbfilename": string;
        "databaseID": string;
        "dbType": string;
        "dbstatus": string;
        "dbenv": string;
        "COCDBID":Number
      }